requirements.txt
torch==2.6.0
transformers==4.55.0
# Pillow==11.3.0
pillow
streamlit==1.26.0
safetensors>=0.6.2
numpy>=1.21.0
